# orgs/core/blackroad-agent-os/schemas/api.yaml
...
openapi: 3.0.3
info:
  title: BlackRoad Agent OS API
  version: 1.0.0
  description: Central controller API for the BlackRoad Agent OS
servers:
  - url: http://localhost:8000
    description: Development server
paths:
...
  /api/agents/{agent_id}:
    get:
      summary: Get specific agent details
      tags: [Agents]
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: 'agent.yaml#/Agent'
        '404':
          description: Agent not found
...
  /api/agents/{agent_id}/workspaces:
    get:
      summary: List workspaces on an agent
      tags: [Agents]
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'agent.yaml#/Workspace'

  # ============ TASKS ============
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []